+++ src/main/java/com/solace/spring/cloud/stream/binder/config/autoconfigure/SolaceTracerConfiguration.java
@@ -5,16 +5,18 @@ import com.solace.spring.cloud.stream.binder.tracing.TracingProxy;
 import io.micrometer.tracing.Tracer;
 import io.micrometer.tracing.propagation.Propagator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.actuate.autoconfigure.tracing.MicrometerTracingAutoConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @ConditionalOnBean({Tracer.class, Propagator.class})
 @Configuration
+@AutoConfiguration(after = MicrometerTracingAutoConfiguration.class)
 public class SolaceTracerConfiguration {
 
     @Bean
-
     public TracingImpl tracingImpl(@Autowired Tracer tracer, @Autowired Propagator propagator) {
         return new TracingImpl(tracer, propagator);
     }